Moves before update query functionality
## What
Moves the before update functionality from yasgui-component.directive.js to sparql view controller.

## Why
At this point this functionality is sparql view specific.

## How

Additional work:
 - removes unnecessary service ontotext-yasgui-web-component.service.js
